Refactoring and export implementation
