+++ cmd/tmsUpload_test.go
@@ -40,13 +40,13 @@ const WRONG_MTA_VERSION = "3.2.1"
 const LAST_CHANGED_AT = "2021-11-16T13:06:05.711Z"
 const INVALID_INPUT_MSG = "Invalid input parameter(s) when getting MTA extension descriptor"
 
-type tmsUploadMockUtils struct {
+type tmsMockUtils struct {
 	*mock.ExecMockRunner
 	*mock.FilesMock
 }
 
-func newTmsUploadTestsUtils() tmsUploadMockUtils {
-	utils := tmsUploadMockUtils{
+func newTmsTestsUtils() tmsMockUtils {
+	utils := tmsMockUtils{
 		ExecMockRunner: &mock.ExecMockRunner{},
 		FilesMock:      &mock.FilesMock{},
 	}
@@ -60,12 +60,14 @@ type communicationInstanceMock struct {
 	uploadMtaExtDescriptorToNodeResponse  tms.MtaExtDescriptor
 	uploadFileResponse                    tms.FileInfo
 	uploadFileToNodeResponse              tms.NodeUploadResponseEntity
+	exportFileToNodeResponse              tms.NodeUploadResponseEntity
 	isErrorOnGetNodes                     bool
 	isErrorOnGetMtaExtDescriptor          bool
 	isErrorOnUpdateMtaExtDescriptor       bool
 	isErrorOnUploadMtaExtDescriptorToNode bool
 	isErrorOnUploadFile                   bool
 	isErrorOnUploadFileToNode             bool
+	isErrorOnExportFileToNode             bool
 }
 
 func (cim *communicationInstanceMock) GetNodes() ([]tms.Node, error) {
@@ -142,6 +144,19 @@ func (cim *communicationInstanceMock) UploadFileToNode(nodeName, fileId, descrip
 	}
 }
 
+func (cim *communicationInstanceMock) ExportFileToNode(nodeName, fileId, description, namedUser string) (tms.NodeUploadResponseEntity, error) {
+	var nodeUploadResponseEntity tms.NodeUploadResponseEntity
+	if description != CUSTOM_DESCRIPTION || nodeName != NODE_NAME || fileId != strconv.FormatInt(FILE_ID, 10) || namedUser != NAMED_USER {
+		return nodeUploadResponseEntity, errors.New(INVALID_INPUT_MSG)
+	}
+
+	if cim.isErrorOnExportFileToNode {
+		return nodeUploadResponseEntity, errors.New("Something went wrong on exporting file to node")
+	} else {
+		return cim.exportFileToNodeResponse, nil
+	}
+}
+
 func mapToJson(m map[string]interface{}) (string, error) {
 	b, err := json.Marshal(m)
 	if err != nil {
@@ -161,7 +176,7 @@ func TestRunTmsUpload(t *testing.T) {
 		fileInfo := tms.FileInfo{Id: FILE_ID, Name: MTA_NAME}
 		communicationInstance := communicationInstanceMock{getNodesResponse: nodes, uploadFileResponse: fileInfo}
 
-		utils := newTmsUploadTestsUtils()
+		utils := newTmsTestsUtils()
 		utils.AddFile(MTA_PATH_LOCAL, []byte("dummy content"))
 
 		mtaYamlBytes, _ := os.ReadFile(MTA_YAML_PATH)
@@ -182,14 +197,14 @@ func TestRunTmsUpload(t *testing.T) {
 		assert.NoError(t, err)
 	})
 
-	t.Run("happy path: no mapping between node nmaes and MTA extension descriptors is provided -> only upload file and upload file to node calls will be executed", func(t *testing.T) {
+	t.Run("happy path: no mapping between node names and MTA extension descriptors is provided -> only upload file and upload file to node calls will be executed", func(t *testing.T) {
 		t.Parallel()
 
 		// init
 		fileInfo := tms.FileInfo{Id: FILE_ID, Name: MTA_NAME}
 		communicationInstance := communicationInstanceMock{uploadFileResponse: fileInfo}
 
-		utils := newTmsUploadTestsUtils()
+		utils := newTmsTestsUtils()
 		utils.AddFile(MTA_PATH_LOCAL, []byte("dummy content"))
 
 		config := tmsUploadOptions{MtaPath: MTA_PATH_LOCAL, CustomDescription: CUSTOM_DESCRIPTION, NamedUser: NAMED_USER, NodeName: NODE_NAME, MtaVersion: MTA_VERSION}
@@ -210,7 +225,7 @@ func TestRunTmsUpload(t *testing.T) {
 		fileInfo := tms.FileInfo{Id: FILE_ID, Name: MTA_NAME}
 		communicationInstance := communicationInstanceMock{getNodesResponse: nodes, getMtaExtDescriptorResponse: mtaExtDescriptor, uploadFileResponse: fileInfo}
 
-		utils := newTmsUploadTestsUtils()
+		utils := newTmsTestsUtils()
 		utils.AddFile(MTA_PATH_LOCAL, []byte("dummy content"))
 
 		mtaYamlBytes, _ := os.ReadFile(MTA_YAML_PATH)
@@ -236,7 +251,7 @@ func TestRunTmsUpload(t *testing.T) {
 
 		// init
 		communicationInstance := communicationInstanceMock{}
-		utils := newTmsUploadTestsUtils()
+		utils := newTmsTestsUtils()
 
 		nodeNameExtDescriptorMapping := map[string]interface{}{NODE_NAME: MTA_EXT_DESCRIPTOR_PATH_LOCAL}
 		nodeNameExtDescriptorMapStr, convErr := mapToJson(nodeNameExtDescriptorMapping)
@@ -255,7 +270,7 @@ func TestRunTmsUpload(t *testing.T) {
 
 		// init
 		communicationInstance := communicationInstanceMock{isErrorOnGetNodes: true}
-		utils := newTmsUploadTestsUtils()
+		utils := newTmsTestsUtils()
 		utils.AddFile(MTA_PATH_LOCAL, []byte("dummy content"))
 
 		nodeNameExtDescriptorMapping := map[string]interface{}{NODE_NAME: MTA_EXT_DESCRIPTOR_PATH_LOCAL}
@@ -276,7 +291,7 @@ func TestRunTmsUpload(t *testing.T) {
 		// init
 		nodes := []tms.Node{{Id: NODE_ID, Name: NODE_NAME}}
 		communicationInstance := communicationInstanceMock{getNodesResponse: nodes}
-		utils := newTmsUploadTestsUtils()
+		utils := newTmsTestsUtils()
 		utils.AddFile(MTA_PATH_LOCAL, []byte("dummy content"))
 
 		nodeNameExtDescriptorMapping := map[string]interface{}{NODE_NAME: MTA_EXT_DESCRIPTOR_PATH_LOCAL}
@@ -297,7 +312,7 @@ func TestRunTmsUpload(t *testing.T) {
 		// init
 		nodes := []tms.Node{{Id: NODE_ID, Name: NODE_NAME}}
 		communicationInstance := communicationInstanceMock{getNodesResponse: nodes}
-		utils := newTmsUploadTestsUtils()
+		utils := newTmsTestsUtils()
 		utils.AddFile(MTA_PATH_LOCAL, []byte("dummy content"))
 
 		mtaYamlBytes, _ := os.ReadFile(INVALID_MTA_YAML_PATH)
@@ -321,7 +336,7 @@ func TestRunTmsUpload(t *testing.T) {
 		// init
 		nodes := []tms.Node{{Id: NODE_ID, Name: NODE_NAME}}
 		communicationInstance := communicationInstanceMock{getNodesResponse: nodes}
-		utils := newTmsUploadTestsUtils()
+		utils := newTmsTestsUtils()
 		utils.AddFile(MTA_PATH_LOCAL, []byte("dummy content"))
 
 		mtaYamlBytes, _ := os.ReadFile(INVALID_MTA_YAML_PATH_2)
@@ -349,7 +364,7 @@ func TestRunTmsUpload(t *testing.T) {
 		// init
 		nodes := []tms.Node{{Id: NODE_ID, Name: NODE_NAME}}
 		communicationInstance := communicationInstanceMock{getNodesResponse: nodes}
-		utils := newTmsUploadTestsUtils()
+		utils := newTmsTestsUtils()
 		utils.AddFile(MTA_PATH_LOCAL, []byte("dummy content"))
 
 		mtaYamlBytes, _ := os.ReadFile(MTA_YAML_PATH)
@@ -391,7 +406,7 @@ func TestRunTmsUpload(t *testing.T) {
 		// init
 		nodes := []tms.Node{{Id: NODE_ID, Name: NODE_NAME}}
 		communicationInstance := communicationInstanceMock{getNodesResponse: nodes, isErrorOnGetMtaExtDescriptor: true}
-		utils := newTmsUploadTestsUtils()
+		utils := newTmsTestsUtils()
 		utils.AddFile(MTA_PATH_LOCAL, []byte("dummy content"))
 
 		mtaYamlBytes, _ := os.ReadFile(MTA_YAML_PATH)
@@ -420,7 +435,7 @@ func TestRunTmsUpload(t *testing.T) {
 		mtaExtDescriptor := tms.MtaExtDescriptor{Id: ID_OF_MTA_EXT_DESCRIPTOR, Description: "Some existing description", MtaId: MTA_ID, MtaExtId: MTA_EXT_ID, MtaVersion: MTA_VERSION, LastChangedAt: LAST_CHANGED_AT}
 		communicationInstance := communicationInstanceMock{getNodesResponse: nodes, getMtaExtDescriptorResponse: mtaExtDescriptor, isErrorOnUpdateMtaExtDescriptor: true}
 
-		utils := newTmsUploadTestsUtils()
+		utils := newTmsTestsUtils()
 		utils.AddFile(MTA_PATH_LOCAL, []byte("dummy content"))
 
 		mtaYamlBytes, _ := os.ReadFile(MTA_YAML_PATH)
@@ -448,7 +463,7 @@ func TestRunTmsUpload(t *testing.T) {
 		nodes := []tms.Node{{Id: NODE_ID, Name: NODE_NAME}}
 		communicationInstance := communicationInstanceMock{getNodesResponse: nodes, isErrorOnUploadMtaExtDescriptorToNode: true}
 
-		utils := newTmsUploadTestsUtils()
+		utils := newTmsTestsUtils()
 		utils.AddFile(MTA_PATH_LOCAL, []byte("dummy content"))
 
 		mtaYamlBytes, _ := os.ReadFile(MTA_YAML_PATH)
@@ -476,7 +491,7 @@ func TestRunTmsUpload(t *testing.T) {
 		nodes := []tms.Node{{Id: NODE_ID, Name: NODE_NAME}}
 		communicationInstance := communicationInstanceMock{getNodesResponse: nodes, isErrorOnUploadFile: true}
 
-		utils := newTmsUploadTestsUtils()
+		utils := newTmsTestsUtils()
 		utils.AddFile(MTA_PATH_LOCAL, []byte("dummy content"))
 
 		mtaYamlBytes, _ := os.ReadFile(MTA_YAML_PATH)
@@ -505,7 +520,7 @@ func TestRunTmsUpload(t *testing.T) {
 		fileInfo := tms.FileInfo{Id: FILE_ID, Name: MTA_NAME}
 		communicationInstance := communicationInstanceMock{getNodesResponse: nodes, uploadFileResponse: fileInfo, isErrorOnUploadFileToNode: true}
 
-		utils := newTmsUploadTestsUtils()
+		utils := newTmsTestsUtils()
 		utils.AddFile(MTA_PATH_LOCAL, []byte("dummy content"))
 
 		mtaYamlBytes, _ := os.ReadFile(MTA_YAML_PATH)
